refactor: Rename EditCategoryBottomSheet to CategoryDetailsEditBottomSheet and remove EditCategoryViewModel
